add separate folders for each approaches
